# internal/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Orchis Foundry API",
    "version": "0.7.0",
    "description": "Self-hosted code platform: git hosting, pull requests, per-user BYO-model supply-chain scanning, and LLM-friendly context endpoints. Authenticate with a personal access token: `Authorization: Bearer orc_pat_<...>`."
  },
  "servers": [{ "url": "/v1" }],
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "orc_pat" }
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
    "/me": {
      "get": { "summary": "Current user", "responses": { "200": { "description": "User" }, "401": { "description": "Not signed in" } } },
      "patch": { "summary": "Update your profile (name, bio)", "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "name": { "type": "string" }, "bio": { "type": "string" } } } } } }, "responses": { "200": { "description": "User" } } }
    },
...
    "/me/tokens/{id}/rotate": { "post": { "summary": "Rotate a token", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "TokenCreated" } } } },
    "/me/tokens/{id}": { "delete": { "summary": "Revoke a token", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "204": { "description": "Deleted" } } } },
...
    "/me/ssh-keys/{id}": { "delete": { "summary": "Remove an SSH key", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "204": { "description": "Deleted" } } } },
...
    "/users/{handle}": { "get": { "summary": "Public user profile (identity + public repos)", "parameters": [{ "name": "handle", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "{handle,name,bio,avatarUrl,memberSince,repos}" }, "404": { "description": "not found" } } } },
...
    "/me/sessions/{id}": { "delete": { "summary": "Revoke a session", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "204": { "description": "Revoked" } } } },
...
    "/repos/{org}/{name}/webhooks": {
      "get": { "summary": "List a repo's webhooks (owner only)", "responses": { "200": { "description": "[Webhook]" } } },
      "post": { "summary": "Create a webhook", "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "url": { "type": "string" }, "events": { "type": "array", "items": { "type": "string", "enum": ["push", "pull_request", "issue", "release", "deploy", "comment"] } }, "secret": { "type": "string" } }, "required": ["url", "events"] } } } }, "responses": { "200": { "description": "Webhook" } } }
    },
    "/repos/{org}/{name}/webhooks/{id}": {
      "patch": { "summary": "Update a webhook (url/events/secret/active)", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "Webhook" } } },
      "delete": { "summary": "Delete a webhook", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "204": { "description": "Deleted" } } }
    },
    "/repos/{org}/{name}/webhooks/{id}/test": { "post": { "summary": "Send a test ping delivery", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "204": { "description": "Queued" } } } },
...
    "/repos/{org}/{name}": {
      "get": { "summary": "Get a repo", "responses": { "200": { "description": "Repo" } } },
      "patch": { "summary": "Update repo settings (rename / description / visibility / default branch)", "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "name": { "type": "string" }, "description": { "type": "string" }, "visibility": { "type": "string", "enum": ["public", "private", "internal"] }, "defaultBranch": { "type": "string" } } } } } }, "responses": { "200": { "description": "Repo" } } },
      "delete": { "summary": "Delete a repo", "responses": { "204": { "description": "Deleted" } } }
    },
...
    "/repos/{org}/{name}/collaborators/{handle}": {
      "put": { "summary": "Add/update a collaborator (admin only)", "parameters": [{ "name": "handle", "in": "path", "required": true, "schema": { "type": "string" } }], "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "role": { "type": "string", "enum": ["read", "write", "admin"] } }, "required": ["role"] } } } }, "responses": { "200": { "description": "{handle,role}" } } },
      "delete": { "summary": "Remove a collaborator (admin only)", "parameters": [{ "name": "handle", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "204": { "description": "Removed" } } }
    },
    "/repos/{org}/{name}/releases": {
      "get": { "summary": "List releases", "responses": { "200": { "description": "[Release]" } } },
      "post": { "summary": "Publish a release (owner only; creates the tag if missing)", "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "tag": { "type": "string" }, "name": { "type": "string" }, "body": { "type": "string" }, "target": { "type": "string" }, "prerelease": { "type": "boolean" } }, "required": ["tag"] } } } }, "responses": { "200": { "description": "Release" } } }
    },
    "/repos/{org}/{name}/releases/{tag}": { "delete": { "summary": "Delete a release + its tag (owner only)", "parameters": [{ "name": "tag", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "204": { "description": "Deleted" } } } },
    "/repos/{org}/{name}/releases/{tag}/tarball": { "get": { "summary": "Download release source as .tar.gz", "parameters": [{ "name": "tag", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "application/gzip" } } } },
    "/repos/{org}/{name}/tree": { "get": { "summary": "File tree (nested)", "parameters": [{ "name": "ref", "in": "query", "schema": { "type": "string" } }], "responses": { "200": { "description": "[Node]" } } } },
    "/repos/{org}/{name}/blob": { "get": { "summary": "File contents", "parameters": [{ "name": "ref", "in": "query", "schema": { "type": "string" } }, { "name": "path", "in": "query", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Blob {content,lang,size,lines}" } } } },
...
    "/repos/{org}/{name}/branches": {
      "get": { "summary": "List branches", "responses": { "200": { "description": "[Branch]" } } },
      "post": { "summary": "Create a branch (owner only)", "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "name": { "type": "string" }, "from": { "type": "string" } }, "required": ["name"] } } } }, "responses": { "200": { "description": "[Branch]" } } }
    },
...
    "/repos/{org}/{name}/commits": {
      "get": { "summary": "List commits", "parameters": [{ "name": "ref", "in": "query", "schema": { "type": "string" } }, { "name": "limit", "in": "query", "schema": { "type": "integer" } }], "responses": { "200": { "description": "[Commit]" } } },
      "post": { "summary": "Commit file edits from the browser (needs write access)", "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "branch": { "type": "string" }, "baseSha": { "type": "string" }, "message": { "type": "string" }, "changes": { "type": "array", "items": { "type": "object", "properties": { "path": { "type": "string" }, "content": { "type": "string" }, "delete": { "type": "boolean" } } } } }, "required": ["branch", "message", "changes"] } } } }, "responses": { "200": { "description": "{sha,branch,created}" } } }
    },
    "/repos/{org}/{name}/commits/draft-message": { "post": { "summary": "Generate a Conventional-Commits message from a diff or changes (uses your model)", "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "diff": { "type": "string" }, "changes": { "type": "array", "items": { "type": "object" } } } } } } }, "responses": { "200": { "description": "{message}" } } } },
    "/repos/{org}/{name}/commits/{sha}": { "get": { "summary": "Commit detail + diff", "parameters": [{ "name": "sha", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "{sha,short,message,author,email,date,additions,deletions,files:[FileDiff]}" } } } },
    "/repos/{org}/{name}/blame": { "get": { "summary": "Per-line authorship for a file", "parameters": [{ "name": "ref", "in": "query", "schema": { "type": "string" } }, { "name": "path", "in": "query", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "[{sha,short,author,when,summary,line,text}]" } } } },
    "/repos/{org}/{name}/outline": { "get": { "summary": "Structural symbols (funcs/classes/types) for a file — AST-aware chunking", "parameters": [{ "name": "ref", "in": "query", "schema": { "type": "string" } }, { "name": "path", "in": "query", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "[{kind,name,lineStart,lineEnd,signature}]" } } } },
    "/repos/{org}/{name}/proposals": {
      "get": { "summary": "List change proposals", "parameters": [{ "name": "status", "in": "query", "schema": { "type": "string", "enum": ["pending", "accepted", "rejected"] } }], "responses": { "200": { "description": "[Proposal]" } } },
      "post": { "summary": "Propose a change (human-in-the-loop gate; needs write access)", "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "title": { "type": "string" }, "summary": { "type": "string" }, "patch": { "type": "string" }, "changes": { "type": "array", "items": { "type": "object" } }, "baseSha": { "type": "string" }, "targetBranch": { "type": "string" } }, "required": ["title"] } } } }, "responses": { "200": { "description": "{id,status}" } } }
    },
    "/repos/{org}/{name}/proposals/{id}": { "get": { "summary": "Proposal detail", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "Proposal" } } } },
    "/repos/{org}/{name}/proposals/{id}/accept": { "post": { "summary": "Accept a proposal: apply to a feature branch + open a PR (needs write access)", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "{status,branch,sha,pullNumber}" } } } },
    "/repos/{org}/{name}/proposals/{id}/reject": { "post": { "summary": "Reject a proposal (needs write access)", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "{status}" } } } },
    "/repos/{org}/{name}/chat": { "post": { "summary": "Chat about the repo with your model; pinned files become context", "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "chatId": { "type": "integer" }, "pinned": { "type": "array", "items": { "type": "string" } }, "messages": { "type": "array", "items": { "type": "object", "properties": { "role": { "type": "string" }, "content": { "type": "string" } } } } }, "required": ["messages"] } } } }, "responses": { "200": { "description": "{chatId,message}" } } } },
...
    "/repos/{org}/{name}/chats/{id}": { "get": { "summary": "Messages in a chat", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "{id,messages}" } } } },
    "/repos/{org}/{name}/compare": { "get": { "summary": "Compare two branches (diff + commits) without a PR", "parameters": [{ "name": "base", "in": "query", "schema": { "type": "string" } }, { "name": "head", "in": "query", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "{base,head,aheadBy,additions,deletions,filesCount,commits,files}" } } } },
    "/repos/{org}/{name}/checks": { "get": { "summary": "Recent check runs across the repo (scan + external CI)", "parameters": [{ "name": "limit", "in": "query", "schema": { "type": "integer" } }], "responses": { "200": { "description": "[Check]" } } } },
    "/repos/{org}/{name}/issues": {
      "get": { "summary": "List issues", "parameters": [{ "name": "state", "in": "query", "schema": { "type": "string", "enum": ["open", "closed", "all"] } }], "responses": { "200": { "description": "[Issue]" } } },
      "post": { "summary": "Open an issue", "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "title": { "type": "string" }, "body": { "type": "string" } }, "required": ["title"] } } } }, "responses": { "200": { "description": "Issue" } } }
    },
    "/repos/{org}/{name}/issues/{num}": {
      "get": { "summary": "Issue detail", "responses": { "200": { "description": "Issue" } } },
      "patch": { "summary": "Edit issue (title/body/state/assignee)", "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "title": { "type": "string" }, "body": { "type": "string" }, "state": { "type": "string", "enum": ["open", "closed"] }, "assignee": { "type": "string" } } } } } }, "responses": { "200": { "description": "Issue" } } }
    },
    "/repos/{org}/{name}/issues/{num}/comments": {
      "get": { "summary": "Issue comments", "responses": { "200": { "description": "[Comment]" } } },
      "post": { "summary": "Comment on an issue", "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "body": { "type": "string" } }, "required": ["body"] } } } }, "responses": { "200": { "description": "Comment" } } }
    },
    "/repos/{org}/{name}/pack": { "get": { "summary": "Whole repo as one markdown document (LLM ingest)", "parameters": [{ "name": "ref", "in": "query", "schema": { "type": "string" } }, { "name": "maxBytes", "in": "query", "schema": { "type": "integer" } }], "responses": { "200": { "description": "text/markdown" } } } },
    "/search/palette": { "get": { "summary": "Command-palette search (repos, PRs, files)", "parameters": [{ "name": "q", "in": "query", "schema": { "type": "string" } }], "responses": { "200": { "description": "{repos:[],pulls:[],files:[]} each item {id,label,sub,route}" } } } },
    "/search/code": { "get": { "summary": "Code search (literal, via git grep)", "parameters": [{ "name": "q", "in": "query", "required": true, "schema": { "type": "string" } }, { "name": "repo", "in": "query", "schema": { "type": "string" } }, { "name": "ref", "in": "query", "schema": { "type": "string" } }, { "name": "lang", "in": "query", "schema": { "type": "string" } }], "responses": { "200": { "description": "[{repo,path,ref,line,text,route}]" } } } },
    "/search/repos": { "get": { "summary": "Repository search", "parameters": [{ "name": "q", "in": "query", "schema": { "type": "string" } }], "responses": { "200": { "description": "[{id,label,sub,route}]" } } } },
    "/search/users": { "get": { "summary": "User search", "parameters": [{ "name": "q", "in": "query", "schema": { "type": "string" } }], "responses": { "200": { "description": "[{id,handle,name,initials,color}]" } } } },
    "/stream": { "get": { "summary": "Server-Sent Events stream (realtime)", "description": "Subscribe to ACL'd topics for live events. Cookie auth only (EventSource cannot set headers). Event kinds: pull.commented, pull.reviewed, pull.merged, push, check.updated.", "parameters": [{ "name": "topics", "in": "query", "description": "Comma-separated: pull:org/name#num, repo:org/name, inbox:<userId>", "schema": { "type": "string" } }], "responses": { "200": { "description": "text/event-stream" } } } },
    "/pulls": { "get": { "summary": "Pull requests across your repos", "parameters": [{ "name": "filter", "in": "query", "schema": { "type": "string", "enum": ["review-requested", "yours", "mentioned", "all"] } }], "responses": { "200": { "description": "[PullSummary]" } } } },
    "/repos/{org}/{name}/pulls": {
      "get": { "summary": "List PRs in a repo", "parameters": [{ "name": "state", "in": "query", "schema": { "type": "string", "enum": ["open", "closed", "merged", "all"] } }], "responses": { "200": { "description": "[PullSummary]" } } },
      "post": { "summary": "Open a PR", "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "title": { "type": "string" }, "head": { "type": "string" }, "base": { "type": "string" }, "body": { "type": "string" } }, "required": ["title", "head"] } } } }, "responses": { "200": { "description": "PullSummary" } } }
    },
    "/repos/{org}/{name}/pulls/{num}": {
      "get": { "summary": "PR detail", "responses": { "200": { "description": "PullSummary" } } },
      "patch": { "summary": "Edit PR title/body", "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "title": { "type": "string" }, "body": { "type": "string" } } } } } }, "responses": { "200": { "description": "PullSummary" } } }
    },
...
    "/repos/{org}/{name}/pulls/{num}/comments": {
      "get": { "summary": "PR comments", "responses": { "200": { "description": "[Comment]" } } },
      "post": { "summary": "Post a comment (top-level or line-anchored)", "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "body": { "type": "string" }, "path": { "type": "string" }, "line": { "type": "integer" }, "side": { "type": "string", "enum": ["left", "right"] } }, "required": ["body"] } } } }, "responses": { "200": { "description": "Comment" } } }
    },
    "/repos/{org}/{name}/pulls/{num}/reviews": { "post": { "summary": "Submit a review", "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "verdict": { "type": "string", "enum": ["comment", "approve", "changes"] }, "body": { "type": "string" } } } } } }, "responses": { "200": { "description": "ok" } } } },
    "/repos/{org}/{name}/pulls/{num}/merge": { "post": { "summary": "Merge a PR", "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "strategy": { "type": "string", "enum": ["merge", "squash", "rebase"] }, "commit_title": { "type": "string" }, "commit_body": { "type": "string" } } } } } }, "responses": { "200": { "description": "merged" }, "409": { "description": "conflict / not open" } } } },
    "/repos/{org}/{name}/pulls/{num}/request-review": { "post": { "summary": "Request a review (owner/author only)", "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "reviewer": { "type": "string" } }, "required": ["reviewer"] } } } }, "responses": { "200": { "description": "ok" } } } },
    "/repos/{org}/{name}/pulls/{num}/request-review/{handle}": { "delete": { "summary": "Remove a requested reviewer", "parameters": [{ "name": "handle", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "204": { "description": "removed" } } } },
...
    "/repos/{org}/{name}/pulls/{num}/summarize": { "post": { "summary": "Summarize the PR with your configured model", "responses": { "200": { "description": "{summary}" }, "409": { "description": "no model configured" } } } },
    "/repos/{org}/{name}/pulls/{num}/explain": { "post": { "summary": "Explain the diff (optional ?file=) with your model", "parameters": [{ "name": "file", "in": "query", "schema": { "type": "string" } }], "responses": { "200": { "description": "{explanation}" } } } },
...
  }
}